+++ src/myfft.h
@@ -17,7 +17,7 @@ namespace myfft
     // The fft is computed on every column of the input matrix.
     // Every input matrix row corresponds to a time evaluation 
     // of a flattened mym x myn matrix (rows concatenated).
-    std::vector<std::vector<densematrix>> fft(densematrix input, long int mym, long int myn);
+    std::vector<std::vector<densematrix>> fft(densematrix input, int mym, int myn);
     
     // Remove the harmonics that are 'threshold' times smaller than the max(abs()) harmonic.
     void removeroundoffnoise(std::vector<std::vector<densematrix>>& input, double threshold = 1e-8);
@@ -27,12 +27,12 @@ namespace myfft
     // each column to a data point of the densematrix inputs. 
     // input[harm][0] gives the values of harmonic 'harm'. 
     // If input[harm] is empty its value is assumed to be 0.
-    densematrix inversefft(std::vector<std::vector<densematrix>>& input, long int numtimevals, long int mym, long int myn);
+    densematrix inversefft(std::vector<std::vector<densematrix>>& input, int numtimevals, int mym, int myn);
         
     // Take a densematrix in the format given by 'inversefft', i.e.
     // with row i corresponding to time evaluation i. Output a 
     // densematrix in which row i corresponds to element i.
-    densematrix toelementrowformat(densematrix timestepsinrows, long int numberofelements);
+    densematrix toelementrowformat(densematrix timestepsinrows, int numberofelements);
     
     // Make sure notsame[i] has the same harmonic content for every i (fill with zeros if not).
     void sameharmonics(std::vector<std::vector<std::vector<densematrix>>>& notsame);
